+++ sigil/storage/adapters/psql/tests/test_psql_seed.py
@@ -4,22 +4,25 @@ import pytest
 
 from sigil.storage.adapters.psql.commands import include_storages
 from sigil.storage.adapters.psql.storages import CampaignStorage
+from sigil.storage.adapters.psql.test_utils import psql_session
 from sigil.storage.seed import seed
 
 
 @pytest.mark.database
-async def test_database_seed(psql_session):
-    async with psql_session.begin():
-        # run the session in a transaction to avoid errors in the seed command
-        campaigns = await CampaignStorage(psql_session).list()
-        assert len(campaigns) == 0
+async def test_database_seed():
+    async with psql_session() as db_session:
+        async with db_session.begin():
+            # run the session in a transaction to avoid errors in the seed command
+            campaigns = await CampaignStorage(db_session).list()
+            assert len(campaigns) == 0
 
-    with mock.patch(
-        "sigil.storage.adapters.psql.commands.async_session", return_value=psql_session
-    ):
-        # redecorate the function to avoid any misconfiguration
-        psql_seed = include_storages(seed.__wrapped__)
-        await psql_seed()
+        with mock.patch(
+            "sigil.storage.adapters.psql.commands.async_session",
+            return_value=db_session,
+        ):
+            # redecorate the function to avoid any misconfiguration
+            psql_seed = include_storages(seed.__wrapped__)
+            await psql_seed()
 
-    campaigns = await CampaignStorage(psql_session).list()
-    assert len(campaigns) > 0
+        campaigns = await CampaignStorage(db_session).list()
+        assert len(campaigns) > 0